# docs/specs/api-contract.yaml
...
openapi: 3.0.3
info:
  title: ORBIT Backend API
  version: 1.0.0
  description: >
    Contract for the ORBIT FastAPI backend. Dashboard and Hermes are both
    clients of this API. The API exposes state and actions only — never rules;
    all conditional/triggered logic lives in n8n. Every state-changing endpoint
    writes an audit_log entry attributed to the calling actor (user | hermes | n8n).
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /auth/login:
    post:
      tags: [auth]
      security: []
      summary: Exchange credentials for an access + refresh token pair
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, format: password }
      responses:
        '200':
          description: Authenticated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '429': { $ref: '#/components/responses/RateLimited' }
  /auth/refresh:
    post:
      tags: [auth]
      security: []
      summary: Rotate a refresh token for a new token pair
      description: The presented refresh token is revoked and replaced (rotation).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        '200':
          description: New token pair
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TokenPair' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/logout:
    post:
      tags: [auth]
      summary: Revoke the current session's refresh token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        '204': { description: Session revoked }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /auth/me:
    get:
      tags: [auth]
      summary: Current authenticated principal
      responses:
        '200':
          description: Principal
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { $ref: '#/components/responses/Unauthorized' }

  # ------------------------------------------------------------- devices ----
...
  /devices/{deviceId}/capabilities:
    parameters:
      - { $ref: '#/components/parameters/DeviceId' }
    get:
      tags: [capabilities]
      summary: List a device's capabilities with current state
      responses:
        '200':
          description: Capabilities
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/CapabilityState' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [capabilities]
      summary: Declare a new capability on a device
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CapabilityCreate' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Capability' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /devices/{deviceId}/capabilities/{capability}:
    parameters:
      - { $ref: '#/components/parameters/DeviceId' }
      - { $ref: '#/components/parameters/CapabilityName' }
    get:
      tags: [capabilities]
      summary: Current state of one capability
      responses:
        '200':
          description: Current state (last retained value)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CapabilityState' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [capabilities]
      summary: Write a value / send a command to a writable capability
      description: >
        Backend validates against the capability's data_type/config, publishes
        to `orbit/devices/{deviceId}/{capability}/set`, and records an audit
        entry attributed to the caller. Fire-and-forget at the MQTT level;
        confirmation arrives via the state topic → WebSocket push.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CommandRequest' }
      responses:
        '202':
          description: Command accepted and published
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CommandAccepted' }
        '403':
          description: Capability is not writable
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: Device is offline
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /devices/{deviceId}/capabilities/{capability}/readings:
    parameters:
      - { $ref: '#/components/parameters/DeviceId' }
      - { $ref: '#/components/parameters/CapabilityName' }
    get:
      tags: [capabilities]
      summary: Historical readings for one capability
      parameters:
        - { name: from, in: query, required: true, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time }, description: Defaults to now }
        - name: interval
          in: query
          description: Optional server-side downsampling bucket (e.g. 1m, 5m, 1h, 1d)
          schema: { type: string, pattern: '^\d+(s|m|h|d)$' }
        - name: aggregate
          in: query
          description: Aggregation when interval is set (numeric capabilities only)
          schema: { type: string, enum: [avg, min, max, last], default: avg }
        - { name: limit, in: query, schema: { type: integer, default: 1000, maximum: 10000 } }
      responses:
        '200':
          description: Readings, ascending by time
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReadingsResponse' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/ValidationError' }

  # --------------------------------------------------------------- rooms ----
...
  /ai/usage:
    get:
      tags: [ai]
      summary: Query AI usage / cost records
      parameters:
        - { name: conversation_id, in: query, schema: { type: string, format: uuid } }
        - { name: model, in: query, schema: { type: string } }
        - { name: from, in: query, schema: { type: string, format: date-time } }
        - { name: to, in: query, schema: { type: string, format: date-time } }
        - name: group_by
          in: query
          description: When set, returns aggregated totals instead of raw rows
          schema: { type: string, enum: [model, day, conversation] }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Offset' }
      responses:
        '200':
          description: Usage rows (or aggregates when group_by is set)
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      oneOf:
                        - $ref: '#/components/schemas/AiUsageEntry'
                        - $ref: '#/components/schemas/AiUsageAggregate'
                  meta: { $ref: '#/components/schemas/PageMeta' }
    post:
      tags: [ai]
      summary: Record AI usage for one model call (Hermes internal)
      description: >
        Write side of the usage log — one row per OpenRouter call, written by
        Hermes via hermes-tool-manifest.json tool `report_ai_usage`
        (internal:true, never exposed to the LLM). Restricted to the hermes
        service principal (actor_type=hermes). `total_tokens` is a generated
        column (prompt + completion) and is not accepted in the body.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AiUsageCreate' }
      responses:
        '201':
          description: Usage row recorded
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AiUsageEntry' }
        '403':
          description: Caller is not the hermes service principal
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
        '422': { $ref: '#/components/responses/ValidationError' }

  # -------------------------------------------------------------- hermes ----
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Users authenticate via /auth/login. Hermes and n8n use long-lived
        service principals (JWTs with actor_type claims 'hermes' / 'n8n')
        issued from backend config — the audit log reads the actor from
        this claim.
  parameters:
    DeviceId:
      name: deviceId
      in: path
      required: true
      schema: { type: string, format: uuid }
    CapabilityName:
      name: capability
      in: path
      required: true
      schema: { type: string, pattern: '^[a-z][a-z0-9_]*$' }
    Limit:
      name: limit
      in: query
      schema: { type: integer, default: 50, minimum: 1, maximum: 500 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, default: 0, minimum: 0 }
  responses:
    Unauthorized:
      description: Missing or invalid credentials/token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Uniqueness or state conflict
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    ValidationError:
      description: Request failed validation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    RateLimited:
      description: Too many attempts
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, example: device_offline }
            message: { type: string }
            details: { type: object, additionalProperties: true }
    PageMeta:
      type: object
      properties:
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
    TokenPair:
      type: object
      required: [access_token, refresh_token, token_type, expires_in]
      properties:
        access_token: { type: string, description: Short-lived JWT }
        refresh_token: { type: string, description: Opaque, rotated on use }
        token_type: { type: string, enum: [bearer] }
        expires_in: { type: integer, description: Access token TTL in seconds }
    User:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        display_name: { type: string }
        role: { type: string, enum: [admin, member] }
...
    Capability:
      type: object
      properties:
        id: { type: string, format: uuid }
        device_id: { type: string, format: uuid }
        capability: { type: string, example: temperature }
        data_type: { type: string, enum: [float, int, bool, string, enum, json] }
        unit: { type: string, nullable: true, example: "°C" }
        access: { type: string, enum: [read, write, read_write] }
        label: { type: string, example: Temperature }
        config: { type: object, additionalProperties: true }
    CapabilityCreate:
      type: object
      required: [capability, data_type, access, label]
      properties:
        capability: { type: string }
        data_type: { type: string, enum: [float, int, bool, string, enum, json] }
        unit: { type: string }
        access: { type: string, enum: [read, write, read_write] }
        label: { type: string }
        config: { type: object, additionalProperties: true }
    CapabilityState:
      allOf:
        - $ref: '#/components/schemas/Capability'
        - type: object
          properties:
            value:
              description: Latest reported value (type per data_type); null if never reported
              nullable: true
            reported_at: { type: string, format: date-time, nullable: true }
    CommandRequest:
      type: object
      required: [value]
      properties:
        value:
          description: Target value; validated against data_type and config
      example: { value: true }
    CommandAccepted:
      type: object
      properties:
        request_id: { type: string, format: uuid }
        published_topic: { type: string, example: "orbit/devices/{id}/power/set" }
        audit_id: { type: string, format: uuid }
    Reading:
      type: object
      properties:
        value: { nullable: true }
        recorded_at: { type: string, format: date-time }
    ReadingsResponse:
      type: object
      properties:
        device_id: { type: string, format: uuid }
        capability: { type: string }
        interval: { type: string, nullable: true }
        aggregate: { type: string, nullable: true }
        data:
          type: array
          items: { $ref: '#/components/schemas/Reading' }
...
    AiUsageEntry:
      type: object
      properties:
        id: { type: string, format: uuid }
        conversation_id: { type: string, format: uuid, nullable: true }
        message_id: { type: string, format: uuid, nullable: true }
        model: { type: string }
        prompt_tokens: { type: integer }
        completion_tokens: { type: integer }
        total_tokens: { type: integer }
        cost_usd: { type: number }
        latency_ms: { type: integer, nullable: true }
        created_at: { type: string, format: date-time }
    AiUsageAggregate:
      type: object
      properties:
        group: { type: string, description: "model slug, ISO date, or conversation id" }
        request_count: { type: integer }
        total_tokens: { type: integer }
        total_cost_usd: { type: number }
    AiUsageCreate:
      type: object
      required: [model, prompt_tokens, completion_tokens, cost_usd]
      properties:
        conversation_id: { type: string, format: uuid, nullable: true }
        message_id: { type: string, format: uuid, nullable: true }
        model: { type: string, description: OpenRouter model slug }
        prompt_tokens: { type: integer, minimum: 0 }
        completion_tokens: { type: integer, minimum: 0 }
        cost_usd: { type: number, minimum: 0 }
        latency_ms: { type: integer, nullable: true }
...